Revert "refactor: 리뷰 반영(접근제어자 등)"
This reverts commit 95b2127bcff4d48c67203d9324d153facc692096.
